[ES|QL] distinguish between trigger kinds in tests (#188604)
## Summary

Part of https://github.com/elastic/kibana/issues/188677

Monaco editor has different [kinds of completion
triggers](https://microsoft.github.io/monaco-editor/typedoc/enums/languages.CompletionTriggerKind.html).
However, the current tests only validate the "TriggerCharacter" events.

This PR prepares the tests to support validating "Invoke" as well.

**Note:** It does change many of the tests from a "TriggerCharacter" to
an "Invoke" scenario. I think this is okay because
- there are still plenty of "TriggerCharacter" tests
- it would take a lot of work to update all the tests
- I will be adding a full set of tests to cover both scenarios as part
of https://github.com/elastic/kibana/issues/188677
- We may rely less and less on trigger characters in the future

---------

(cherry picked from commit 76c6f550dc5c945f1410bb53d54923c7363c7660)

# Conflicts:
#	packages/kbn-esql-validation-autocomplete/src/autocomplete/__tests__/helpers.ts
#	packages/kbn-esql-validation-autocomplete/src/autocomplete/autocomplete.test.ts
